time type fixed

+++ philo_one/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: hpark <hpark@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/12 13:47:00 by hpark             #+#    #+#             */
-/*   Updated: 2020/08/18 13:19:33 by hpark            ###   ########.fr       */
+/*   Updated: 2020/08/19 16:57:44 by hpark            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,10 +22,10 @@
 typedef struct	s_vars{
 	int				n_philo;
 	unsigned long	t_die;
-	int				t_eat;
-	int				t_sleep;
+	unsigned long	t_eat;
+	unsigned long	t_sleep;
 	int				n_must_eat;
-	int				t_start;
+	unsigned long	t_start;
 	int				n_done;
 	int				died;
 	pthread_mutex_t	*fork;
